requirements.txt
streamlit
google-generativeai
requests
openai
PyPDF2
python-dotenv
python-jose
reportlab
pymupdf